refactor: merge login api to one only

+++ src/main/java/com/dutact/web/auth/dto/ConfirmDto.java
@@ -13,5 +13,5 @@ public class ConfirmDto {
     @NotBlank
     private String email;
     @NotNull
-    private Integer otp;
+    private Long otp;
 }
