+++ src/main/java/com/dutact/web/auth/dto/student/StudentConfirmResetPasswordDto.java
@@ -15,5 +15,5 @@ public class StudentConfirmResetPasswordDto {
     private String newPassword;
 
     @NotNull
-    private Integer otp;
+    private Long otp;
 }
